Documents: Refactor utils package
