+++ gcampus/documents/document.py
@@ -20,16 +20,54 @@ __all__ = [
     "as_file",
 ]
 
+import mimetypes
+import os
 import pathlib
+import posixpath
 from io import BytesIO
 from typing import Optional, List, Union
+from urllib.parse import urlsplit
 
+from django.conf import settings
+from django.contrib.staticfiles import finders
 from django.http import HttpRequest
 from django.template.loader import render_to_string
 from weasyprint import HTML, Document
+from weasyprint.urls import URLFetchingError, default_url_fetcher
 
 from gcampus.core import get_base_url
-from gcampus.documents.utils import mock_request, url_fetcher
+from gcampus.documents.utils import mock_request
+
+
+def url_fetcher(url: str, **kwargs):
+    try:
+        o = urlsplit(url)
+    except ValueError:
+        return default_url_fetcher(url, **kwargs)
+    if o.netloc == settings.PRIMARY_HOST and o.path.startswith(settings.STATIC_URL):
+        path = os.path.relpath(o.path, settings.STATIC_URL)
+        normalized_path = posixpath.normpath(path).lstrip("/")
+        mime_type, encoding = mimetypes.guess_type(path, strict=True)
+        if getattr(settings, "AWS_STATIC_USE_STATIC_URL", False):
+            from django.contrib.staticfiles.storage import staticfiles_storage
+
+            file_obj = staticfiles_storage.open(normalized_path)
+        else:
+            absolute_path = finders.find(normalized_path)
+            if not absolute_path:
+                raise FileNotFoundError()
+            if not os.path.isfile(absolute_path):
+                raise URLFetchingError(
+                    f"File '{url}' (resolved to '{absolute_path}') not found!"
+                )
+            file_obj = open(absolute_path, "rb")
+        return {
+            "mime_type": mime_type,
+            "encoding": encoding,
+            "file_obj": file_obj,
+        }
+    else:
+        return default_url_fetcher(url, **kwargs)
 
 
 def render_document_template(
